[all] Nasty copy/paste bug on Radio Timer (inactivity alarm raises far too quickly!) [Taranis] Bad display of selected lines over bitmaps in MODELSEL screen

+++ src/opentx.h
@@ -914,7 +914,7 @@ extern uint16_t lastMixerDuration;
   #define MAXTRACE (LCD_W - 8)
   extern uint8_t s_traceBuf[MAXTRACE];
   extern uint8_t s_traceWr;
-  extern int8_t s_traceCnt;
+  extern int     s_traceCnt;
 #endif
 
 #if defined(PCBTARANIS)
